+++ src/main/java/act/test/util/ScenarioComparator.java
@@ -9,9 +9,9 @@ package act.test.util;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -47,6 +47,11 @@ public class ScenarioComparator implements Comparator<Scenario> {
             } else if (p1 && !p2) {
                 return 1;
             }
+            if (o1.ignore && !o2.ignore) {
+                return -1;
+            } else if (o2.ignore) {
+                return 1;
+            }
         }
         List<Scenario> d1 = depends(o1, new ArrayList<Scenario>());
         List<Scenario> d2 = depends(o2, new ArrayList<Scenario>());
@@ -59,6 +64,7 @@ public class ScenarioComparator implements Comparator<Scenario> {
         int n = o1.partition.compareTo(o2.partition);
         return 0 != n ? n : o1.name.compareTo(o2.name);
     }
+
     private List<Scenario> depends(Scenario s, List<Scenario> depends) {
         for (String name : s.depends) {
             Scenario scenario = scenarioManager.get(name);